fix(daemon, core): tighten gh_poll error recovery and storage cleanup
- list_jobs failure now logs + continues to next run instead of aborting
  the entire repo pass
- get_annotations and get_log_tail failures now log via eprintln! instead
  of silently dropping (all three marked TODO for structured logging)
- list_runs gets .with_context() for clearer error attribution
- insert_log_excerpt: truncated cast aligned to as i32 (matches storage.rs
  convention at lines 375/380/384)
- RunRow/JobRow get doc comments distinguishing them from gh_api types
- ?13 param comment in upsert_run warns about paired first/last_seen_at

+++ crates/hippo-core/src/storage.rs
@@ -2001,6 +2001,8 @@ pub mod workflow_store {
 
     use crate::gh_annotations::parse as parse_annotation;
 
+    /// Storage-layer projection of a workflow_run for upsert.
+    /// Distinct from `hippo_daemon::gh_api::WorkflowRun` (the API response type).
     pub struct RunRow<'a> {
         pub id: i64,
         pub repo: &'a str,
@@ -2027,6 +2029,8 @@ pub mod workflow_store {
                 status=excluded.status, conclusion=excluded.conclusion,
                 completed_at=excluded.completed_at, last_seen_at=excluded.last_seen_at,
                 raw_json=excluded.raw_json",
+            // `?13` is bound once for both first_seen_at (initial insert) and
+            // last_seen_at (update). If you reorder params, keep them paired.
             params![
                 run.id,
                 run.repo,
@@ -2046,6 +2050,8 @@ pub mod workflow_store {
         Ok(())
     }
 
+    /// Storage-layer projection of a workflow_job for upsert.
+    /// Distinct from `hippo_daemon::gh_api::Job` (the API response type).
     pub struct JobRow<'a> {
         pub id: i64,
         pub run_id: i64,
@@ -2111,7 +2117,7 @@ pub mod workflow_store {
         conn.execute(
             "INSERT INTO workflow_log_excerpts (job_id, step_name, excerpt, truncated)
              VALUES (?1,?2,?3,?4)",
-            params![job_id, step_name, excerpt, truncated as i64],
+            params![job_id, step_name, excerpt, truncated as i32],
         )?;
         Ok(())
     }
